# contracts/api/v2/openapi.json
{
  "components": {
    "schemas": {
...
      "OnboardingFlow": {
        "additionalProperties": false,
        "properties": {
          "steps": {
            "description": "Ordered steps for the surface and profile",
            "items": {
              "$ref": "#/components/schemas/OnboardingStep"
            },
            "type": "array"
          },
          "tour_id": {
            "description": "The current tour; progress is recorded against it",
            "examples": [
              "core-2026-07"
            ],
            "type": "string"
          },
          "version": {
            "description": "Flow format version",
            "examples": [
              1
            ],
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "tour_id",
          "steps"
        ],
        "type": "object"
      },
      "OnboardingProgressInputBody": {
        "additionalProperties": false,
        "properties": {
          "completed": {
            "description": "Mark the tour completed",
            "examples": [
              false
            ],
            "type": "boolean"
          },
          "last_step": {
            "description": "The step the profile reached",
            "examples": [
              "playback-quality"
            ],
            "maxLength": 64,
            "type": "string"
          },
          "skipped": {
            "description": "Mark the tour skipped",
            "examples": [
              false
            ],
            "type": "boolean"
          },
          "tour_id": {
            "description": "The tour the progress belongs to; a tour other than the current one is refused",
            "examples": [
              "core-2026-07"
            ],
            "maxLength": 64,
            "type": "string"
          }
        },
        "type": "object"
      },
      "OnboardingSetting": {
        "additionalProperties": false,
        "properties": {
          "control": {
            "description": "Control kind: segmented, toggle, or select",
            "examples": [
              "segmented"
            ],
            "type": "string"
          },
          "default": {
            "description": "Preselected value",
            "examples": [
              "auto"
            ],
            "type": "string"
          },
          "key": {
            "description": "The profile member or setting key written",
            "examples": [
              "quality_preference"
            ],
            "type": "string"
          },
          "label": {
            "description": "Annotation of a toggle control",
            "examples": [
              "Show subtitles"
            ],
            "type": "string"
          },
          "options": {
            "description": "Choices; empty for a toggle",
            "items": {
              "$ref": "#/components/schemas/OnboardingSettingOption"
            },
            "type": "array"
          },
          "target": {
            "description": "Which API the choice writes through: profile_field (updateProfile), setting, or device_setting",
            "examples": [
              "profile_field"
            ],
            "type": "string"
          }
        },
        "required": [
          "target",
          "key",
          "control",
          "options"
        ],
        "type": "object"
      },
      "OnboardingSettingOption": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "description": "Display label",
            "examples": [
              "Automatic"
            ],
            "type": "string"
          },
          "value": {
            "description": "The value written when chosen",
            "examples": [
              "auto"
            ],
            "type": "string"
          }
        },
        "required": [
          "value",
          "label"
        ],
        "type": "object"
      },
      "OnboardingState": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "description": "RFC 3339 instant the tour was completed; absent otherwise",
            "examples": [
              "2026-01-02T03:04:05Z"
            ],
            "type": "string"
          },
          "done": {
            "description": "True once the tour was completed or skipped: do not show it again",
            "examples": [
              false
            ],
            "type": "boolean"
          },
          "last_step": {
            "description": "The last step the profile reached; absent before the tour starts",
            "examples": [
              "playback-quality"
            ],
            "type": "string"
          },
          "skipped_at": {
            "description": "RFC 3339 instant the tour was skipped; absent otherwise",
            "examples": [
              "2026-01-02T03:04:05Z"
            ],
            "type": "string"
          },
          "tour_id": {
            "description": "The current tour",
            "examples": [
              "core-2026-07"
            ],
            "type": "string"
          }
        },
        "required": [
          "tour_id",
          "done"
        ],
        "type": "object"
      },
      "OnboardingStep": {
        "additionalProperties": false,
        "properties": {
          "action_label": {
            "description": "Label of the optional feature_card action",
            "examples": [
              "Browse requests"
            ],
            "type": "string"
          },
          "body": {
            "examples": [
              "A quick tour of what you can do."
            ],
            "type": "string"
          },
          "id": {
            "description": "Stable step id; the value recordOnboardingProgress takes as last_step",
            "examples": [
              "welcome"
            ],
            "type": "string"
          },
          "illustration": {
            "description": "Client-side asset key; never a URL",
            "examples": [
              "welcome"
            ],
            "type": "string"
          },
          "kind": {
            "description": "Step kind: welcome, feature_card, setting_choice, or handoff; clients skip unknown kinds",
            "examples": [
              "welcome"
            ],
            "type": "string"
          },
          "links": {
            "description": "Outbound links; empty when none",
            "items": {
              "$ref": "#/components/schemas/OnboardingStepLink"
            },
            "type": "array"
          },
          "route": {
            "description": "Client route for a feature_card action or a handoff",
            "examples": [
              "/requests"
            ],
            "type": "string"
          },
          "setting": {
            "$ref": "#/components/schemas/OnboardingSetting",
            "description": "Present on setting_choice steps"
          },
          "title": {
            "examples": [
              "Welcome to Silo"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "links"
        ],
        "type": "object"
      },
      "OnboardingStepLink": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "description": "Button label",
            "examples": [
              "App Store"
            ],
            "type": "string"
          },
          "url": {
            "description": "External URL",
            "examples": [
              "https://apps.apple.com/app/silo"
            ],
            "type": "string"
          }
        },
        "required": [
          "label",
          "url"
        ],
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "description": "Safe occurrence-specific explanation; not for control flow",
            "examples": [
              "The request did not pass validation; see errors."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Field-level validation details",
            "items": {
              "$ref": "#/components/schemas/ProblemError"
            },
            "type": "array"
          },
          "instance": {
            "description": "urn:silo:request:\u003crequest-id\u003e, matching the X-Request-ID response header",
            "examples": [
              "urn:silo:request:000000000000000000000003"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code, equal to the response status",
            "examples": [
              422
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "Short summary fixed for the problem type",
            "examples": [
              "Validation failed"
            ],
            "type": "string"
          },
          "type": {
            "description": "Stable problem type URI; the final segment is the problem identifier",
            "examples": [
              "https://siloserver.org/docs/api/v2/problems/validation_failed"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "type": "object"
      },
      "ProblemError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Stable machine-readable code for the failure",
            "examples": [
              "required"
            ],
            "type": "string"
          },
          "detail": {
            "description": "Safe human-readable explanation; never the rejected value",
            "examples": [
              "expected required property name to be present"
            ],
            "type": "string"
          },
          "location": {
            "description": "Where the error occurred: body.*, query.*, path.* or header.*",
            "examples": [
              "body.name"
            ],
            "type": "string"
          }
        },
        "required": [
          "location",
          "code",
          "detail"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "A session token or an API key in the Authorization header. The server decides which it holds.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Silo API",
    "version": "2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/onboarding/flow": {
      "get": {
        "operationId": "getOnboardingFlow",
        "parameters": [
          {
            "description": "The household profile acting for this request; it must belong to the authenticated account.",
            "in": "header",
            "name": "X-Profile-Id",
            "required": true,
            "schema": {
              "examples": [
                "1"
              ],
              "type": "string"
            }
          },
          {
            "description": "Verification proof for a PIN-locked profile, issued by POST /api/v1/profiles/{id}/verify-pin until that operation moves to v2; required only when the declared profile is locked",
            "in": "header",
            "name": "X-Profile-Token",
            "schema": {
              "examples": [
                "pvt_5f3a9c1e7b2d4e8fa0c6"
              ],
              "type": "string"
            }
          },
          {
            "description": "The client surface; web when absent",
            "explode": false,
            "in": "query",
            "name": "surface",
            "schema": {
              "description": "The client surface; web when absent",
              "enum": [
                "web",
                "phone",
                "tv"
              ],
              "examples": [
                "tv"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingFlow"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get the first-run tour for the acting profile on one surface.",
        "tags": [
          "onboarding"
        ],
        "x-silo-class": "profile_scoped",
        "x-silo-service-backed": true
      }
    },
    "/api/v2/onboarding/progress": {
      "post": {
        "operationId": "recordOnboardingProgress",
        "parameters": [
          {
            "description": "The household profile acting for this request; it must belong to the authenticated account.",
            "in": "header",
            "name": "X-Profile-Id",
            "required": true,
            "schema": {
              "examples": [
                "1"
              ],
              "type": "string"
            }
          },
          {
            "description": "Verification proof for a PIN-locked profile, issued by POST /api/v1/profiles/{id}/verify-pin until that operation moves to v2; required only when the declared profile is locked",
            "in": "header",
            "name": "X-Profile-Token",
            "schema": {
              "examples": [
                "pvt_5f3a9c1e7b2d4e8fa0c6"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingProgressInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "408": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Timeout"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Record the acting profile's progress through the current tour.",
        "tags": [
          "onboarding"
        ],
        "x-silo-class": "profile_scoped",
        "x-silo-service-backed": true
      }
    },
    "/api/v2/onboarding/state": {
      "get": {
        "operationId": "getOnboardingState",
        "parameters": [
          {
            "description": "The household profile acting for this request; it must belong to the authenticated account.",
            "in": "header",
            "name": "X-Profile-Id",
            "required": true,
            "schema": {
              "examples": [
                "1"
              ],
              "type": "string"
            }
          },
          {
            "description": "Verification proof for a PIN-locked profile, issued by POST /api/v1/profiles/{id}/verify-pin until that operation moves to v2; required only when the declared profile is locked",
            "in": "header",
            "name": "X-Profile-Token",
            "schema": {
              "examples": [
                "pvt_5f3a9c1e7b2d4e8fa0c6"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingState"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get the acting profile's progress through the current tour.",
        "tags": [
          "onboarding"
        ],
        "x-silo-class": "profile_scoped",
        "x-silo-service-backed": true
      }
    },
...
  }
}